requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
google-cloud-storage==2.14.0
boto3==1.34.34
python-json-logger==2.0.7

